backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pandas==2.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
